requirements.txt
Django==1.11
django-filter==1.0.4
djangorestframework==3.6.3
gunicorn==19.7.1
Markdown==2.6.8
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
